requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
tavily-python==0.3.0
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
python-multipart==0.0.6
google-generativeai==0.8.3
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.24.3 